requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.5
anyio==3.3.4
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
curlify==2.2.1
databases==0.5.3
dataclasses-json==0.5.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
facebook-business==12.0.1
facebook-sdk==3.1.0
fastapi==0.70.0
fastapi-contrib==0.2.11
fastapi-utils==0.2.1
fb==0.4.0
forex-python==1.8
frozenlist==1.2.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
intel-openmp==2021.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jaeger-client==4.8.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
llvmlite==0.37.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2021.4.0
motor==2.5.1
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numba==0.54.1
numexpr==2.8.0
numpy==1.20.3
oauthlib==3.1.1
opentracing==2.4.0
orjson==3.6.4
ormar==0.10.22
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psycopg==3.0.4
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pycountry==20.7.3
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-facebook-api==0.12.3
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.9
starlette==0.16.0
tbb==2021.4.0
terminado==0.12.1
testpath==0.5.0
threadloop==1.0.2
threadpoolctl==3.0.0
thrift==0.15.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-inspect==0.7.1
typing_extensions==4.0.0
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
widgetsnbextension==3.5.2
yarl==1.7.2
